+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.65 (Pending):
 * Remove excessive logging in GreenMailTestModule.
 * Add AbstractSharedTest.postInjector() hook method that is invoked after the injector is created.
+* Import ValueUtil from all downstream projects.
 
 ## 0.64:
 * Issue restart to GlassFish container after jms host changes.